TASK: Simplify value object handling and serialization
